% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscor.R
\name{tscor}
\alias{tscor}
\title{Compute correlation matrix of an n by t time series matrix}
\usage{
tscor(x)
}
\arguments{
\item{x}{An n by t time series matrix}
}
\value{
An n by n correlation matrix
}
\description{
Compute correlation matrix of an n by t time series matrix
}
