% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscor.R
\name{tscorplot}
\alias{tscorplot}
\title{Return a visualization of time series correlation matrix}
\usage{
tscorplot(x)
}
\arguments{
\item{x}{An n by t time series matrix}
}
\value{
A heatmap visualizing correlation matrix
}
\description{
Return a visualization of time series correlation matrix
}
