% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscor.R
\name{tsplot}
\alias{tsplot}
\title{Return a visualization of time series}
\usage{
tsplot(x)
}
\arguments{
\item{x}{An n by t time series matrix}
}
\value{
A gglot with n time series on one graph
}
\description{
Return a visualization of time series
}
